Use repr of value for DeviceArray.__repr__
Currently, DevieArray.__repr__ returns the class name and the size of
the object. This is inconsistent with np.ndarray.__repr__ which produces
code to reconstruct the array (including the contents, unless the
contents are too long to print cleanly). This provides a strange
reminder that we aren't using "real" numpy so it may be nicer to just
use the standard numpy __repr__ instead.