Add underscore prefixes to xla.py methods. Add some docstrings.
This PR does not take an opinion about which methods should be public, instead it attempts to codify the world as it is now by adding underscore prefixes to methods that are local to xla.py. It is frequently helpful to me to be able to identify which functions are API entry points and which are not.