Add rmsprop_momentum optimizer (same as TF RMSProp)
The TensorFlow RMSProp optimizer supports an additional momentum
parameter, which allows adding momentum to the RMSProp update.
Having momentum requires keeping around additional state, which might
not be desirable when using the standard RMSProp optimizer, so I've
created an additional optimizer for this case.
RMSProp with momentum can be necessary to reproduce some research papers.