Change jax.numpy scalar types to return 0D JAX arrays when instantiated. (#1836)
* Change jax.numpy scalar types to return 0D JAX arrays rather than NumPy scalars when instantiated.

jax.numpy and numpy have slightly different promotion behaviors. For consistency with JAX arrays, we would like the result of, say, `jax.numpy.int32(7)` to have the same promotion behavior as `jax.numpy.array(7, dtype=jax.numpy.int32)`. The easiest way to do this is to have the jax.numpy scalars return 0D arrays when instantiated; the difference between NumPy scalars and arrays is not a fundamental one and we do not need to distinguish between them in JAX.