[jax2tf] Fix min_max and add_mul harnesses.
Some of the array parameters were wrapped in StaticArg(), which
resulted in the function not being actually converted through
jax2tf.

One oddity is that the test now requires "always_custom_assert" to
be set to True to run properly, implying that even in compiled
mode, the results produced by jax and jax2tf are inconsistent.