Ensure that check_jaxpr is done with abstract values
Prior to this it was possible, e.g., for code that contains a Literal,
such as  to result in FLOPS during checking.

The assertion is broken by many tests unless we raise_to_shape for Literals.

I have timed the checks on my laptop and I do not see a reduction in the
total test time.