Remove tracing special case for trivial subcomputations.
This undoes #3169 now that omnistaging is in. I'm having trouble
reproducing the original issue with omnistaging disabled (#3143), but
my understanding is that we should no longer have issues with values
being forced by jitted subcomputations, since jit no longer partially
evaluates.

The motivation for this change is that this special case occasionally
causes confusion and unexpected behavior. For example, it bypasses
call-specific error checking (#5188) and can cause trivial nested
pmaps and sharded_jits, which are useful for pre-sharding data across
devices, to have no effect.