clear some caches when setting up jaxpr typecheck tests
In order to test that the typechecker identifies invalid jaxprs, some
tests modify jaxprs in place. This is typically not allowed, since
jaxprs are assumed immutable, and may be cached. As a workaround, this
change clears the relevant caches before every test. This ought to
prevent some order-dependent test failures.