Restore support for nested xmaps when using the (standard) replica lowering
Fortunately this wasn't too difficult, as most of the code was already
there. The biggest issues were a lack of axis name substitution and
inability to handle multiple resource assignments for a single logical
axis.

I'm planning to add a more comprehensive test suite for this, but I'd
like to wait for the pdot test PR (#5459) to land first. It contains a
bunch of utilities that would come in handy.