Parse and verify pjit in/out_axis_resources
This lets us focus the handling of all the quirks of `PartitionSpec` in
one place and should make it easier to extend support to xmap-like specs
in the future. This also adds new checks for repeated assignment of mesh
axes which would previously went unnoticed (the first annotation would
simply get ignored).

PiperOrigin-RevId: 370425384